recipe_value_system/requirements.txt
SQLAlchemy>=2.0.0
redis>=4.5.0
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
pyyaml>=6.0.0
pytest>=7.3.1
pytest-cov>=4.1.0
mypy>=1.5.1
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
typing-extensions>=4.7.1
dataclasses>=0.6
python-slugify>=8.0.1

# Data export dependencies
pandas>=2.0.0
pyarrow>=14.0.1
xlsxwriter>=3.1.0
openpyxl>=3.1.2
pyexcel>=0.7.0
pyexcel-xlsx>=0.6.0
pyexcel-ods>=0.6.0
tabulate>=0.9.0
python-pptx>=0.6.21
reportlab>=4.0.4
markdown2>=2.4.10
jinja2>=3.1.2
fastapi>=0.109.0
uvicorn>=0.27.0
click>=8.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Additional data export formats
avro-python3>=1.10.2
msgpack>=1.0.5
feather-format>=0.4.1
hdf5plugin>=4.1.1
tables>=3.8.0
orc>=0.3.0
apache-beam>=2.46.0
docx>=0.2.4
python-docx>=0.8.11
latex>=0.7.0
aspose-words>=23.12.0
protobuf>=4.21.12
apache-flink>=1.17.1
neo4j>=5.13.0
pymongo>=4.6.1
